api/requirements.txt
pandas
pandas_ta
flask
flask_restful
python-binance
ta
matplotlib
flask-cors
